Restructure pagination dir; tanspile and add typings
